requirements.txt
flask
flask-bcrypt
python-dotenv
pyjwt
requests
google-generativeai
sqlitecloud
gunicorn
